Refactor list projection
