+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/types/semtype/FixedLengthArray.java
@@ -58,4 +58,8 @@ public record FixedLengthArray(SemType[] initial, int fixedLength) {
     public static FixedLengthArray empty() {
         return EMPTY;
     }
+
+    public FixedLengthArray shallowCopy() {
+        return new FixedLengthArray(initial.clone(), fixedLength);
+    }
 }
